refactor(helpers): 입력 헬퍼 함수 분리
